Fixed Bladestorm:
Now you can use disarms and bleed effects against a warrior in bladestorm.
NOTE: Don`t know where is the real source of problem. Make a special case for bladestorm on IsImmuneToSpell() and IsImmuneToSpellEffect() isn`t the correct solution, we must find why a warrior is immune to that effect, when the immunity to this effects is added to m_spellImmune list.